Change the NumericIdProvider to use `u64` IDs
Instead of providing a guarantee that we can convert between
`usize` and `u64` we make the assumptions that it's unlikely
that we're running on an architecture larger than 64-bits and
we use a `u64` directly.

+++ pubsub/src/manager.rs
@@ -68,10 +68,10 @@ impl NumericIdProvider {
 }
 
 impl IdProvider for NumericIdProvider {
-	type Id = usize;
+	type Id = u64;
 
 	fn next_id(&self) -> Self::Id {
-		self.current_id.fetch_add(1, Ordering::AcqRel)
+		self.current_id.fetch_add(1, Ordering::AcqRel) as u64
 	}
 }
